+++ apps/studio/components/interfaces/Account/AccessTokens/NewAccessTokenSheet.tsx
@@ -1,0 +1,208 @@
+import { zodResolver } from '@hookform/resolvers/zod'
+import { ExternalLink } from 'lucide-react'
+import Link from 'next/link'
+import { useState } from 'react'
+import { type SubmitHandler, useForm } from 'react-hook-form'
+import { toast } from 'sonner'
+import { z } from 'zod'
+
+import { useAccessTokenCreateMutation } from 'data/access-tokens/access-tokens-create-mutation'
+import {
+  Sheet,
+  SheetContent,
+  SheetFooter,
+  SheetHeader,
+  SheetTitle,
+  Button,
+  Form_Shadcn_,
+  ScrollArea,
+  Separator,
+} from 'ui'
+import { Admonition } from 'ui-patterns'
+
+import { TokenBasicInfoForm } from './Form/TokenBasicInfoForm'
+import { TokenResourceAccessForm } from './Form/TokenResourceAccessForm'
+import { TokenPermissionsForm } from './Form/TokenPermissionsForm'
+
+const PermissionRowSchema = z.object({
+  resource: z.string().min(1, 'Please select a resource'),
+  action: z.string().min(1, 'Please select an action'),
+})
+
+const TokenSchema = z.object({
+  tokenName: z.string().min(1, 'Please enter a name for the token'),
+  expirationDate: z.enum(['No expiry', '7 days', '30 days', '90 days', '180 days', 'Custom']),
+  resourceAccess: z.enum(['all-orgs', 'selected-orgs', 'selected-projects']),
+  selectedOrganizations: z.array(z.string()).optional(),
+  selectedProjects: z.array(z.string()).optional(),
+  organizationPermissions: z.record(z.string(), z.string()).optional(),
+  projectPermissions: z.record(z.string(), z.string()).optional(),
+  permissionRows: z.array(PermissionRowSchema).optional(),
+})
+
+export interface NewAccessTokenSheetProps {
+  visible: boolean
+  onOpenChange: (open: boolean) => void
+  tokenScope: 'V0' | undefined
+  onCreateToken: (token: any) => void
+}
+
+const NewAccessTokenSheet = ({ 
+  visible, 
+  onOpenChange, 
+  tokenScope, 
+  onCreateToken 
+}: NewAccessTokenSheetProps) => {
+  const [resourceSearchOpen, setResourceSearchOpen] = useState(false)
+
+  const form = useForm<z.infer<typeof TokenSchema>>({
+    resolver: zodResolver(TokenSchema),
+    defaultValues: {
+      tokenName: '',
+      expirationDate: 'No expiry',
+      resourceAccess: 'all-orgs',
+      selectedOrganizations: [],
+      selectedProjects: [],
+      organizationPermissions: {},
+      projectPermissions: {},
+      permissionRows: [],
+    },
+    mode: 'onChange',
+  })
+  const { mutate: createAccessToken, isLoading } = useAccessTokenCreateMutation()
+
+  const resourceAccess = form.watch('resourceAccess')
+  const expirationDate = form.watch('expirationDate')
+  const permissionRows = form.watch('permissionRows') || []
+
+  const onSubmit: SubmitHandler<z.infer<typeof TokenSchema>> = async (values) => {
+    // Validate that at least one permission is configured
+    if (!permissionRows || permissionRows.length === 0) {
+      toast.error('Please configure at least one permission.')
+      return
+    }
+
+    // Validate that all permission rows have both resource and action
+    const hasValidPermissions = permissionRows.every((row) => row.resource && row.action)
+    if (!hasValidPermissions) {
+      toast.error('Please ensure all permissions have both resource and action selected.')
+      return
+    }
+
+    createAccessToken(
+      {
+        name: values.tokenName,
+        scope: tokenScope,
+      },
+      {
+        onSuccess: (data) => {
+          toast.success('Access token created successfully')
+          onCreateToken(data)
+          handleClose()
+        },
+      }
+    )
+  }
+
+  const handleClose = () => {
+    form.reset({
+      tokenName: '',
+      expirationDate: 'No expiry',
+      resourceAccess: 'all-orgs',
+      selectedOrganizations: [],
+      selectedProjects: [],
+      organizationPermissions: {},
+      projectPermissions: {},
+      permissionRows: [],
+    })
+    onOpenChange(false)
+  }
+
+  return (
+    <Sheet
+      open={visible}
+      onOpenChange={(open) => {
+        if (!open) handleClose()
+        onOpenChange(open)
+      }}
+    >
+      <SheetContent
+        showClose={false}
+        size="default"
+        className="!min-w-[600px] flex flex-col h-full gap-0"
+      >
+        <SheetHeader>
+          <SheetTitle>
+            {tokenScope === 'V0' ? 'Generate token for experimental API' : 'Generate New Token'}
+          </SheetTitle>
+        </SheetHeader>
+        <ScrollArea className="flex-1 max-h-[calc(100vh-116px)]">
+          <div className="flex flex-col overflow-visible">
+            {tokenScope === 'V0' && (
+              <div className="px-4 sm:px-5 py-4 pb-4">
+                <Admonition
+                  type="warning"
+                  title="The experimental API provides additional endpoints which allows you to manage your organizations and projects."
+                  description={
+                    <>
+                      <p>
+                        These include deleting organizations and projects which cannot be undone.
+                        As such, be very careful when using this API.
+                      </p>
+                      <div className="mt-4">
+                        <Button asChild type="default" icon={<ExternalLink />}>
+                          <Link
+                            href="https://api.supabase.com/api/v0"
+                            target="_blank"
+                            rel="noreferrer"
+                          >
+                            Experimental API documentation
+                          </Link>
+                        </Button>
+                      </div>
+                    </>
+                  }
+                />
+              </div>
+            )}
+
+            <Form_Shadcn_ {...form}>
+              <div className="flex flex-col gap-0 overflow-visible">
+                {/* Basic Information Section */}
+                <TokenBasicInfoForm control={form.control} expirationDate={expirationDate} />
+
+                <Separator />
+
+                {/* Resource Access Section */}
+                <TokenResourceAccessForm control={form.control} resourceAccess={resourceAccess} />
+
+                <Separator />
+
+                {/* Permissions Section */}
+                <TokenPermissionsForm
+                  control={form.control}
+                  setValue={form.setValue}
+                  watch={form.watch}
+                  resourceSearchOpen={resourceSearchOpen}
+                  setResourceSearchOpen={setResourceSearchOpen}
+                />
+              </div>
+            </Form_Shadcn_>
+          </div>
+        </ScrollArea>
+        <SheetFooter className="!justify-end w-full mt-auto py-4 border-t">
+          <div className="flex gap-2">
+            <Button type="default" disabled={isLoading} onClick={handleClose}>
+              Cancel
+            </Button>
+            <Button onClick={form.handleSubmit(onSubmit)} loading={isLoading}>
+              Generate token
+            </Button>
+          </div>
+        </SheetFooter>
+      </SheetContent>
+    </Sheet>
+  )
+}
+
+export default NewAccessTokenSheet
